Fix test failed after test_reset_password if user use own password
If use set his own password of FIRST_SUPERUSER_PASSWORD in .env file, after run test_reset_password case, the password will be changed to "changethis" becasue it hardcorded in this testcase. 
Following testcases will all be failed because of wrong password.
Change it to reset to password in .env file.